orchestrator_api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
celery
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyyaml
httpx
openai
regex
rapidfuzz
langdetect
reportlab
jinja2
python-dotenv
cryptography
